Milestone 2a: A/K/F accrual and settlement math
Pure, independently-testable libraries for the lazy A/K/F engine core, ported
from the Rust engine and re-expressed in seconds for L1.

- Accrual.sol (spec §5.3 + §1.7): staircaseNext clamps a raw oracle target toward
  the effective engine price by at most floor(pLast*maxPriceMoveBpsPerSec*dt/1e4);
  accrue enforces the per-second price-move envelope before any mutation, then does
  mark-to-market (K_long += A_long*ΔP, K_short -= A_short*ΔP, per side with OI) and
  funding (F_long -= A_long*total, F_short += A_short*total) and finalizes
  pLast/fundPxLast/slotLast. Mutates only the two side accumulators — O(1) per
  market, the property that makes no-continuous-crank viable on L1.
- Settlement.sol (spec §5.1/§5.2): effectivePosQ = sign(basis)*floor(|basis|*A/aBasis)
  (zero on epoch mismatch), kfPnlDelta = floor->-inf of the combined K/F numerator,
  riskNotionalOf.
- FixedPointMath.divFloorSigned: floor-toward-negative-infinity signed division so
  losses round in the vault's favor.
- Types: widen K/F and account economic fields to int256/uint256 (EVM-native word;
  removes the i128 overflow pressure) and add the two-bucket warmup reserve fields.
- EngineHarness exposes the storage-based Accrual and memory-based Settlement for
  isolated unit tests, mirroring the Rust engine's pure unit tests.

DESIGN.md: documented the margining unit convention (engine uses linear quote-
denominated PnL; coin-margining is the vault-token choice; the SDK's /oracle form is
display only) and the int256 K/F width decision.

forge clean && forge test green: 81 tests, 0 failed
(+25: Accrual 14, Settlement 11).

Trade/liquidate wiring continues in 2b/2c.

+++ src/PerpMarket.sol
@@ -97,8 +97,8 @@ contract PerpMarket is ReentrancyGuard {
         factory = msg.sender;
 
         // Each side's A starts at ADL_ONE ("1.0").
-        g.longSide.a = uint128(Constants.ADL_ONE);
-        g.shortSide.a = uint128(Constants.ADL_ONE);
+        g.longSide.a = Constants.ADL_ONE;
+        g.shortSide.a = Constants.ADL_ONE;
         g.slotLast = uint64(block.timestamp);
 
         emit Initialized(cfg.collateralToken, admin_);
@@ -182,15 +182,14 @@ contract PerpMarket is ReentrancyGuard {
         token.safeTransferFrom(msg.sender, address(this), amount);
         uint256 credited = token.balanceOf(address(this)) - balBefore;
         if (credited == 0) revert ZeroAmount();
-        // Guard the uint128 capital cast (and the per-account checked add). The spec's
-        // engine-units MAX_VAULT_TVL cap is reintroduced in Milestone 2 once token ->
-        // engine-unit price normalization exists; raw 18-decimal token wei is not 1:1
-        // with engine quote-atom units, so it must not be compared against that bound here.
+        // Conservative per-deposit sanity bound. The spec's engine-units MAX_VAULT_TVL
+        // cap is reintroduced once token -> engine-unit price normalization exists
+        // (raw 18-decimal token wei is not 1:1 with engine quote-atom units).
         if (credited > type(uint128).max) revert AmountTooLarge();
 
         // effects
         Types.Account storage a = accounts[positionId];
-        a.capital += uint128(credited);
+        a.capital += credited;
         g.cTot += credited;
         g.vault += credited;
 
@@ -231,7 +230,7 @@ contract PerpMarket is ReentrancyGuard {
         }
 
         // effects (CEI)
-        a.capital -= uint128(amount);
+        a.capital -= amount;
         g.cTot -= amount;
         g.vault -= amount;
